@@ -26,6 +26,7 @@
 
 using System.Collections.Generic;
 using System.IO;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using Crypter.API.Attributes;
@@ -39,6 +40,7 @@ using EasyMonads;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Timeouts;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Crypter.API.Controllers;
@@ -53,13 +55,23 @@ public class FileTransferController : TransferControllerBase
     {
         _sender = sender;
     }
-
+    
+    /// <summary>
+    /// The ideal way to upload a file of any size.
+    /// However, the endpoints to receive chunked files is preferred while this Chromium issue remains unaddressed:
+    /// https://issues.chromium.org/issues/339788214
+    /// </summary>
+    /// <param name="username"></param>
+    /// <param name="request"></param>
+    /// <returns></returns>
     [HttpPost]
     [MaybeAuthorize]
+    [RequestFormLimits(MultipartBodyLengthLimit = long.MaxValue)]
+    [RequestTimeout(300000)]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UploadTransferResponse))]
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ErrorResponse))]
     public async Task<IActionResult> UploadFileTransferAsync([FromQuery] string? username,
-        [FromForm] UploadFileTransferReceipt request)
+        [FromForm] UploadFileTransferBundle request)
     {
         Maybe<string> maybeUsername = string.IsNullOrEmpty(username)
             ? Maybe<string>.None
@@ -76,6 +88,79 @@ public class FileTransferController : TransferControllerBase
                 neither: MakeErrorResponse(UploadTransferError.UnknownError));
     }
 
+    [HttpPost("multipart/initialize")]
+    [Authorize]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UploadTransferResponse))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ErrorResponse))]
+    public async Task<IActionResult> InitializeMultipartFileTransferAsync([FromQuery] string? username,
+        [FromBody] UploadFileTransferRequest request)
+    {
+        Maybe<string> maybeUsername = string.IsNullOrEmpty(username)
+            ? Maybe<string>.None
+            : username;
+
+        InitializeMultipartFileTransferCommand command = new InitializeMultipartFileTransferCommand(UserId, maybeUsername, request);
+        return await _sender.Send(command)
+            .MatchAsync(
+                left: MakeErrorResponse,
+                right: Ok,
+                neither: MakeErrorResponse(UploadTransferError.UnknownError));
+    }
+    
+    [HttpPost("multipart/upload")]
+    [Authorize]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UploadTransferResponse))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ErrorResponse))]
+    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ErrorResponse))]
+    public async Task<IActionResult> UploadMultipartFileTransferAsync([FromQuery] string id, [FromQuery] int position,
+        [FromForm] IFormFile? ciphertext)
+    {
+        SaveMultipartFileTransferCommand command = new SaveMultipartFileTransferCommand(UserId, id, position, ciphertext?.OpenReadStream());
+        return await _sender.Send(command)
+            .MatchAsync(
+                left: MakeErrorResponse,
+                right: _ => Accepted(),
+                neither: MakeErrorResponse(UploadMultipartFileTransferError.UnknownError));
+            
+        IActionResult MakeErrorResponse(UploadMultipartFileTransferError error)
+        {
+#pragma warning disable CS8524
+            return error switch
+            {
+                UploadMultipartFileTransferError.UnknownError => MakeErrorResponseBase(HttpStatusCode.InternalServerError, error),
+                UploadMultipartFileTransferError.NotFound => MakeErrorResponseBase(HttpStatusCode.NotFound, error),
+                UploadMultipartFileTransferError.AggregateTooLarge
+                    or UploadMultipartFileTransferError.OutOfSpace=> MakeErrorResponseBase(HttpStatusCode.BadRequest, error)
+            };
+#pragma warning restore CS8524
+        }
+    }
+    
+    [HttpPost("multipart/finalize")]
+    [Authorize]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UploadTransferResponse))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ErrorResponse))]
+    public async Task<IActionResult> FinalizeMultipartFileTransferAsync([FromQuery] string id)
+    {
+        FinalizeMultipartFileTransferCommand command = new FinalizeMultipartFileTransferCommand(UserId, id);
+        return await _sender.Send(command)
+            .MatchAsync(
+                left: MakeErrorResponse,
+                right: _ => Accepted(),
+                neither: MakeErrorResponse(FinalizeMultipartFileTransferError.UnknownError));
+        
+        IActionResult MakeErrorResponse(FinalizeMultipartFileTransferError error)
+        {
+#pragma warning disable CS8524
+            return error switch
+            {
+                FinalizeMultipartFileTransferError.UnknownError => MakeErrorResponseBase(HttpStatusCode.InternalServerError, error),
+                FinalizeMultipartFileTransferError.NotFound => MakeErrorResponseBase(HttpStatusCode.NotFound, error)
+            };
+#pragma warning restore CS8524
+        }
+    }
+    
     [HttpGet("received")]
     [Authorize]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<UserReceivedFileDTO>))]
@@ -127,6 +212,7 @@ public class FileTransferController : TransferControllerBase
     }
 
     [HttpGet("ciphertext/anonymous")]
+    [RequestTimeout(300000)]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(FileStreamResult))]
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ErrorResponse))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ErrorResponse))]
@@ -146,6 +232,7 @@ public class FileTransferController : TransferControllerBase
 
     [HttpGet("ciphertext/user")]
     [MaybeAuthorize]
+    [RequestTimeout(300000)]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(FileStreamResult))]
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ErrorResponse))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ErrorResponse))]
@@ -162,5 +249,30 @@ public class FileTransferController : TransferControllerBase
                 MakeErrorResponse,
                 x => new FileStreamResult(x, "application/octet-stream"),
                 MakeErrorResponse(DownloadTransferCiphertextError.UnknownError));
+    }
+
+    [HttpPost("multipart/abandon")]
+    [Authorize]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UploadTransferResponse))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ErrorResponse))]
+    public async Task<IActionResult> AbandonMultipartFileTransferAsync([FromQuery] string id)
+    {
+        AbandonMultipartFileTransferCommand command = new AbandonMultipartFileTransferCommand(UserId, id);
+        return await _sender.Send(command)
+            .MatchAsync(
+                left: MakeErrorResponse,
+                right: _ => Accepted(),
+                neither: MakeErrorResponse(AbandonMultipartFileTransferError.UnknownError));
+        
+        IActionResult MakeErrorResponse(AbandonMultipartFileTransferError error)
+        {
+#pragma warning disable CS8524
+            return error switch
+            {
+                AbandonMultipartFileTransferError.UnknownError => MakeErrorResponseBase(HttpStatusCode.InternalServerError, error),
+                AbandonMultipartFileTransferError.NotFound => MakeErrorResponseBase(HttpStatusCode.NotFound, error)
+            };
+#pragma warning restore CS8524
+        }
     }
 }
