@@ -59,7 +59,7 @@ public class MessageTransferController : TransferControllerBase
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UploadTransferResponse))]
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ErrorResponse))]
     public async Task<IActionResult> UploadMessageTransferAsync([FromQuery] string? username,
-        [FromForm] UploadMessageTransferReceipt request)
+        [FromForm] UploadMessageTransferBundle request)
     {
         Maybe<string> maybeUsername = string.IsNullOrEmpty(username)
             ? Maybe<string>.None
