@@ -28,7 +28,6 @@ using System;
 using System.IO;
 using Crypter.Common.Client.Interfaces.HttpClients;
 using Crypter.Common.Client.Transfer.Models;
-using Crypter.Common.Enums;
 using Crypter.Crypto.Common;
 using Crypter.Crypto.Common.KeyExchange;
 using Crypter.Crypto.Common.StreamEncryption;
@@ -42,19 +41,16 @@ public class UploadHandler : IUserUploadHandler
     protected readonly ICryptoProvider CryptoProvider;
     protected readonly ClientTransferSettings ClientTransferSettings;
 
-    protected TransferUserType TransferUserType = TransferUserType.Anonymous;
-
     protected int ExpirationHours;
 
     protected bool SenderDefined;
 
-    protected byte[] KeyExchangeNonce;
+    protected readonly byte[] KeyExchangeNonce;
     protected Maybe<byte[]> SenderPrivateKey = Maybe<byte[]>.None;
 
     protected Maybe<string> RecipientUsername = Maybe<string>.None;
     protected Maybe<byte[]> RecipientKeySeed = Maybe<byte[]>.None;
-
-    protected Maybe<byte[]> RecipientPrivateKey = Maybe<byte[]>.None;
+    
     protected Maybe<byte[]> RecipientPublicKey = Maybe<byte[]>.None;
 
     protected UploadHandler(ICrypterApiClient crypterApiClient, ICryptoProvider cryptoProvider,
@@ -70,13 +66,11 @@ public class UploadHandler : IUserUploadHandler
     public void SetSenderInfo(byte[] privateKey)
     {
         SenderDefined = true;
-        TransferUserType = TransferUserType.User;
         SenderPrivateKey = privateKey;
     }
 
     public void SetRecipientInfo(string username, byte[] publicKey)
     {
-        TransferUserType = TransferUserType.User;
         RecipientUsername = username;
         RecipientPublicKey = publicKey;
     }
@@ -92,11 +86,10 @@ public class UploadHandler : IUserUploadHandler
         Span<byte> seed = CryptoProvider.Random.GenerateRandomBytes((int)CryptoProvider.KeyExchange.SeedSize);
         RecipientKeySeed = seed.ToArray();
         X25519KeyPair recipientKeyPair = CryptoProvider.KeyExchange.GenerateKeyPairDeterministic(seed);
-        RecipientPrivateKey = recipientKeyPair.PrivateKey;
         RecipientPublicKey = recipientKeyPair.PublicKey;
     }
 
-    protected (Func<EncryptionStream> encryptionStreamOpener, byte[]? senderPublicKey, byte[] proof) GetEncryptionInfo(
+    protected (Func<Action<double>?, EncryptionStream> encryptionStreamOpener, byte[]? senderPublicKey, byte[] proof) GetEncryptionInfo(
         Func<Stream> plaintextStreamOpener, long streamSize)
     {
         if (RecipientUsername.IsNone)
@@ -127,8 +120,14 @@ public class UploadHandler : IUserUploadHandler
 
         return (EncryptionStreamOpener, senderPublicKeyToUpload, proof);
 
-        EncryptionStream EncryptionStreamOpener()
-            => new EncryptionStream(plaintextStreamOpener(), streamSize, encryptionKey,
-                CryptoProvider.StreamEncryptionFactory, ClientTransferSettings.MaxReadSize, ClientTransferSettings.PadSize);
+        EncryptionStream EncryptionStreamOpener(Action<double>? updateCallback = null)
+            => new EncryptionStream(
+                plaintextStreamOpener,
+                streamSize, 
+                encryptionKey,
+                CryptoProvider.StreamEncryptionFactory,
+                ClientTransferSettings.MaxReadSize,
+                ClientTransferSettings.PadSize,
+                updateCallback);
     }
 }
