@@ -25,13 +25,19 @@
  */
 
 using System;
+using System.Buffers;
+using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
+using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using Crypter.Common.Client.Interfaces.HttpClients;
 using Crypter.Common.Client.Transfer.Handlers.Base;
 using Crypter.Common.Client.Transfer.Models;
 using Crypter.Common.Contracts.Features.Transfer;
 using Crypter.Common.Enums;
+using Crypter.Common.Mappers;
 using Crypter.Crypto.Common;
 using Crypter.Crypto.Common.StreamEncryption;
 using EasyMonads;
@@ -52,8 +58,7 @@ public class UploadFileHandler : UploadHandler
     {
     }
 
-    internal void SetTransferInfo(Func<Stream> fileStreamOpener, string fileName, long fileSize, string fileContentType,
-        int expirationHours)
+    internal void SetTransferInfo(Func<Stream> fileStreamOpener, string fileName, long fileSize, string fileContentType, int expirationHours)
     {
         _fileStreamOpener = fileStreamOpener;
         _fileName = fileName;
@@ -63,7 +68,7 @@ public class UploadFileHandler : UploadHandler
         _transferInfoSet = true;
     }
 
-    public Task<Either<UploadTransferError, UploadHandlerResponse>> UploadAsync()
+    public Task<Either<UploadTransferError, UploadHandlerResponse>> UploadAsync(Action<double>? updateCallback = null)
     {
         if (!_transferInfoSet)
         {
@@ -71,14 +76,131 @@ public class UploadFileHandler : UploadHandler
                 .From(UploadTransferError.UnknownError)
                 .AsTask();
         }
-        
-        (Func<EncryptionStream> encryptionStreamOpener, byte[]? senderPublicKey, byte[] proof) = GetEncryptionInfo(_fileStreamOpener!, _fileSize);
+
+        (Func<Action<double>?, EncryptionStream> encryptionStreamOpener, byte[]? senderPublicKey, byte[] proof) = GetEncryptionInfo(_fileStreamOpener!, _fileSize);
         UploadFileTransferRequest request = new UploadFileTransferRequest(_fileName!, _fileContentType!, senderPublicKey,
             KeyExchangeNonce, proof, ExpirationHours);
-        return CrypterApiClient.FileTransfer
-            .UploadFileTransferAsync(RecipientUsername, request, encryptionStreamOpener, SenderDefined)
-            .MapAsync<UploadTransferError, UploadTransferResponse, UploadHandlerResponse>(x =>
-                new UploadHandlerResponse(x.HashId, ExpirationHours, TransferItemType.File, x.UserType,
-                    RecipientKeySeed));
+        
+        if (SenderDefined)
+        {
+            // Initialize
+            return CrypterApiClient.FileTransfer
+                // Initialize
+                .InitializeMultipartFileTransferAsync(RecipientUsername, request)
+
+                // Upload
+                .BindAsync(async initializeResult =>
+                {
+                    Dictionary<int, Task<Either<UploadMultipartFileTransferError, Unit>>> indexedUploadResults = new Dictionary<int, Task<Either<UploadMultipartFileTransferError, Unit>>>();
+                    EncryptionStream encryptionStream = encryptionStreamOpener(updateCallback);
+                    IAsyncEnumerable<Func<MemoryStream>> asyncEnumerable = SplitEncryptionStreamAsync(encryptionStream);
+
+                    ParallelOptions parallelOptions = new ParallelOptions
+                    {
+                        MaxDegreeOfParallelism = ClientTransferSettings.MaximumMultipartParallelism
+                    };
+                    SemaphoreSlim uploadLock = new SemaphoreSlim(1);
+                    bool fault = false;
+                    int currentPosition = 0;
+                    await Parallel.ForEachAsync(asyncEnumerable, parallelOptions, async (streamOpener, _) =>
+                    {
+                        if (!fault)
+                        {
+                            try
+                            {
+                                await uploadLock.WaitAsync(CancellationToken.None);
+                                Task<Either<UploadMultipartFileTransferError, Unit>> uploadTask = CrypterApiClient.FileTransfer
+                                    .UploadMultipartFileTransferAsync(initializeResult.HashId, currentPosition, streamOpener)
+                                    .ContinueWith(x =>
+                                    {
+                                        if (!x.Result.IsRight)
+                                        {
+                                            fault = true;
+                                        }
+
+                                        return x.Result;
+                                    }, CancellationToken.None);
+                                indexedUploadResults.Add(currentPosition, uploadTask);
+                                currentPosition++;
+                                uploadLock.Release();
+                                await uploadTask;
+                            }
+                            catch (Exception)
+                            {
+                                fault = true;
+                                throw;
+                            }
+                        }
+                    });
+
+                    Either<UploadMultipartFileTransferError, Unit> errorOrSuccess = indexedUploadResults
+                        .OrderBy(x => x.Key)
+                        .Select(x => x.Value.Result)
+                        .FirstOrDefault(x => x.IsLeft || x.IsNeither, Either<UploadMultipartFileTransferError, Unit>.FromRight(Unit.Default));
+                    
+                    return await errorOrSuccess
+                        .MapLeft(error => error.ConvertToUploadTransferError())
+                        .BindAsync(async _ => await CrypterApiClient.FileTransfer.FinalizeMultipartFileTransferAsync(initializeResult.HashId)
+                            .MapLeftAsync<FinalizeMultipartFileTransferError, Unit, UploadTransferError>(error => error.ConvertToUploadTransferError()))
+                        .MapAsync<UploadTransferError, Unit, UploadHandlerResponse>(_ => new UploadHandlerResponse(initializeResult.HashId, ExpirationHours, TransferItemType.File, initializeResult.TransferUserType, RecipientKeySeed))
+                        .DoLeftOrNeitherAsync(
+                            leftAsync: async _ => await CrypterApiClient.FileTransfer.AbandonMultipartFileTransferAsync(initializeResult.HashId),
+                            neitherAsync: async () => await CrypterApiClient.FileTransfer.AbandonMultipartFileTransferAsync(initializeResult.HashId));
+                });
+        }
+        else
+        {
+            return CrypterApiClient.FileTransfer
+                .UploadFileTransferAsync(RecipientUsername, request, encryptionStreamOpener, SenderDefined, updateCallback)
+                .MapAsync<UploadTransferError, UploadTransferResponse, UploadHandlerResponse>(x =>
+                    new UploadHandlerResponse(x.HashId, ExpirationHours, TransferItemType.File, x.UserType,
+                        RecipientKeySeed));
+        }
+        
+        async IAsyncEnumerable<Func<MemoryStream>> SplitEncryptionStreamAsync(EncryptionStream encryptionStream)
+        {
+            bool endOfStream = false;
+            Stopwatch loopStopwatch = new Stopwatch();
+            short blocksPerRequest = ClientTransferSettings.InitialMultipartReadBlocks;
+            do
+            {
+                loopStopwatch.Restart();
+                int bufferSize = blocksPerRequest * encryptionStream.MinimumBufferSize;
+                byte[] buffer = ArrayPool<byte>.Shared.Rent(bufferSize);
+
+                int totalBytesRead = 0;
+                for (int i = 0; i < blocksPerRequest; i++)
+                {
+                    int bytesRead = await encryptionStream.ReadAsync(buffer.AsMemory(totalBytesRead, encryptionStream.MinimumBufferSize));
+                    totalBytesRead += bytesRead;
+                    if (bytesRead == 0)
+                    {
+                        endOfStream = true;
+                        break;
+                    }
+                }
+                
+                if (totalBytesRead > 0)
+                {
+                    try
+                    {
+                        yield return () => new MemoryStream(buffer, 0, totalBytesRead);
+                    }
+                    finally
+                    {
+                        ArrayPool<byte>.Shared.Return(buffer, clearArray: true);   
+                    }
+                }
+
+                if (loopStopwatch.Elapsed < TimeSpan.FromSeconds(1) && blocksPerRequest < ClientTransferSettings.MaximumMultipartReadBlocks)
+                {
+                    blocksPerRequest += 5;
+                }
+                else if (loopStopwatch.Elapsed > TimeSpan.FromSeconds(1) && blocksPerRequest > ClientTransferSettings.InitialMultipartReadBlocks)
+                {
+                    blocksPerRequest -= 5;
+                }
+            } while (!endOfStream);
+        }
     }
 }
