@@ -28,7 +28,43 @@ namespace Crypter.Common.Client.Transfer.Models;
 
 public class ClientTransferSettings
 {
-    public short MaximumTransferSizeMiB { get; init; }
+    /// <summary>
+    /// Set the limit for the maximum request body size when uploading files using a buffer.
+    /// </summary>
+    public short MaximumUploadBufferSizeMB { get; init; }
+    
+    /// <summary>
+    /// Set the limit for the maximum request body size when uploading files using request streaming.
+    /// </summary>
+    public short MaximumUploadStreamSizeMB { get; init; }
+    
+    /// <summary>
+    /// Set the limit for the maximum file size when uploading files using multiple requests.
+    /// </summary>
+    public short MaximumMultipartUploadSizeMB { get; init; }
+    
+    /// <summary>
+    /// Set the limit for the number of blocks in a single multipart request body.
+    /// </summary>
+    public short MaximumMultipartReadBlocks { get; init; }
+    
+    /// <summary>
+    /// Set the initial number of read blocks in a single multipart request body.
+    /// </summary>
+    public short InitialMultipartReadBlocks { get; init; }
+    
+    /// <summary>
+    /// Set the maximum degrees of parallelism for multipart uploads.
+    /// </summary>
+    public int MaximumMultipartParallelism { get; init; }
+    
+    /// <summary>
+    /// Set the number of bytes of plaintext to read at one time when encrypting a file or message.
+    /// </summary>
     public int MaxReadSize { get; init; }
+    
+    /// <summary>
+    /// Set the maximum number of bytes to use as padding when encrypting a file or message.
+    /// </summary>
     public int PadSize { get; init; }
 }
