@@ -87,7 +87,7 @@ internal class GetUserMessageCiphertextCommandHandler
         return await GetFileStreamAsync(itemId.Value, nullableRequesterUserId, request.Proof)
             .DoRightAsync(async _ =>
             {
-                SuccessfulTransferDownloadEvent downloadEvent = new SuccessfulTransferDownloadEvent(itemId.Value, TransferItemType.Message, TransferUserType.User, nullableRequesterUserId, !_transferHasRecipient, DateTimeOffset.UtcNow);
+                SuccessfulTransferDownloadEvent downloadEvent = new SuccessfulTransferDownloadEvent(itemId.Value, TransferItemType.Message, TransferUserType.User, nullableRequesterUserId, _transferHasRecipient, DateTimeOffset.UtcNow);
                 await _publisher.Publish(downloadEvent, CancellationToken.None);
             })
             .DoLeftOrNeitherAsync(
@@ -127,7 +127,7 @@ internal class GetUserMessageCiphertextCommandHandler
             return DownloadTransferCiphertextError.InvalidRecipientProof;
         }
         
-        _transferHasRecipient = !databaseData.RecipientId.HasValue;
+        _transferHasRecipient = databaseData.RecipientId.HasValue;
         return _transferRepository
             .GetTransfer(itemId, TransferItemType.Message, TransferUserType.User, !_transferHasRecipient)
             .ToEither(DownloadTransferCiphertextError.NotFound);
