@@ -1,5 +1,5 @@
 ﻿@*
- * Copyright (C) 2023 Crypter File Transfer
+ * Copyright (C) 2024 Crypter File Transfer
  *
  * This file is part of the Crypter file transfer project.
  *
@@ -78,6 +78,6 @@
     </div>
     <div class="mt-3 container text-center">
         <div>Code licensed <a class="text-decoration-none" href="https://www.gnu.org/licenses/agpl-3.0.en.html" target="_blank">AGPLv3</a></div>
-        <div>&copy; 2023 Crypter File Transfer</div>
+        <div>&copy; 2024 Crypter File Transfer</div>
     </div>
 </footer>
