@@ -50,7 +50,7 @@
                             }
                         </li>
                         <li class="list-group-item">
-                            <span class="fw-bold">Encrypted Size: </span>@(ByteSize.FromBytes(_messageSize).ToBinaryString())
+                            <span class="fw-bold">Encrypted Size: </span>@(ByteSize.FromBytes(_messageSize).ToString())
                         </li>
                         <li class="list-group-item">
                             <span class="fw-bold">Created: </span>@Created.ToString("MM/dd/yyyy h:mm tt")
