@@ -26,27 +26,36 @@
 
 using System;
 using System.IO;
+using System.Runtime.Versioning;
 using System.Threading.Tasks;
+using Crypter.Common.Client.Enums;
 using Crypter.Common.Client.Transfer.Handlers;
 using Crypter.Common.Client.Transfer.Models;
 using Crypter.Common.Contracts.Features.Transfer;
+using Crypter.Web.Services;
 using EasyMonads;
 using Microsoft.AspNetCore.Components.Forms;
 
 namespace Crypter.Web.Shared.Transfer;
 
+[SupportedOSPlatform("browser")]
 public partial class UploadFileTransfer : IDisposable
 {
     private IBrowserFile? _selectedFile;
     
-    private long _maxFileSizeBytes = 0;
+    private long _maxBufferSizeMB = 0;
+    private long _maxStreamSizeMB = 0;
+    private long _maxMultipartSizeMB = 0;
+    
     private string _dropClass = string.Empty;
     private const string DropzoneDrag = "dropzone-drag";
     private const string NoFileSelected = "No file selected.";
 
     protected override void OnInitialized()
     {
-        _maxFileSizeBytes = UploadSettings.MaximumTransferSizeMiB * (long)Math.Pow(2, 20);
+        _maxBufferSizeMB = ClientTransferSettings.MaximumUploadBufferSizeMB * Convert.ToInt64(Math.Pow(10, 6));
+        _maxStreamSizeMB = ClientTransferSettings.MaximumUploadStreamSizeMB * Convert.ToInt64(Math.Pow(10, 6));
+        _maxMultipartSizeMB = ClientTransferSettings.MaximumMultipartUploadSizeMB * Convert.ToInt64(Math.Pow(10, 6));
     }
 
     private void HandleDragEnter()
@@ -59,40 +68,76 @@ public partial class UploadFileTransfer : IDisposable
         _dropClass = string.Empty;
     }
 
-    private void HandleFileInputChange(InputFileChangeEventArgs e)
+    private async Task HandleFileInputChangeAsync(InputFileChangeEventArgs e)
     {
         _dropClass = string.Empty;
         ErrorMessage = string.Empty;
 
         IBrowserFile file = e.File;
-
-        if (file.Size > _maxFileSizeBytes)
+        
+        if (await UserSessionService.IsLoggedInAsync())
         {
-            ErrorMessage = $"The max file size is {UploadSettings.MaximumTransferSizeMiB} MB.";
-            return;
+            TransmissionType = TransferTransmissionType.Multipart;
+        }
+        else if (BrowserFunctions.BrowserSupportsRequestStreaming())
+        {
+            TransmissionType = TransferTransmissionType.Stream;
         }
 
-        _selectedFile = file;
+        // ReSharper disable once SwitchStatementHandlesSomeKnownEnumValuesWithDefault
+        switch (TransmissionType)
+        {
+            case TransferTransmissionType.Buffer when file.Size > _maxBufferSizeMB:
+                ErrorMessage = $"The max file size is {ClientTransferSettings.MaximumUploadBufferSizeMB} MB. Login for file sizes up to {ClientTransferSettings.MaximumMultipartUploadSizeMB} MB.";
+                break;
+            case TransferTransmissionType.Stream when file.Size > _maxStreamSizeMB:
+                ErrorMessage = $"The max file size is {ClientTransferSettings.MaximumUploadStreamSizeMB} MB.";
+                break;
+            case TransferTransmissionType.Multipart when file.Size > _maxMultipartSizeMB:
+                ErrorMessage = $"The max file size is {ClientTransferSettings.MaximumMultipartUploadSizeMB} MB.";
+                break;
+            default:
+                _selectedFile = file;
+                break;
+        }
+        
+        StateHasChanged();
     }
 
-    protected async Task OnEncryptClicked()
+    private async Task OnEncryptClicked()
     {
         if (_selectedFile is null)
         {
             ErrorMessage = NoFileSelected;
             return;
         }
-
+        
         EncryptionInProgress = true;
         ErrorMessage = string.Empty;
-
-        await SetProgressMessage("Encrypting file");
+        
+        await SetProgressMessageAsync("Encrypting file");
 
         UploadFileHandler fileUploader = TransferHandlerFactory.CreateUploadFileHandler(FileStreamOpener,
             _selectedFile.Name, _selectedFile.Size, _selectedFile.ContentType, ExpirationHours);
 
         SetHandlerUserInfo(fileUploader);
-        Either<UploadTransferError, UploadHandlerResponse> uploadResponse = await fileUploader.UploadAsync();
+
+#pragma warning disable CS8524
+        Action<double>? progressUpdater = TransmissionType switch
+        {
+            TransferTransmissionType.Buffer => null,
+            TransferTransmissionType.Stream
+                or TransferTransmissionType.Multipart => SetUploadPercentage
+        };
+#pragma warning restore CS8524
+
+        Either<UploadTransferError, UploadHandlerResponse> uploadResponse = await fileUploader.UploadAsync(progressUpdater);
+
+        if (TransmissionType is TransferTransmissionType.Stream or TransferTransmissionType.Multipart)
+        {
+            await Task.Delay(400);
+        }
+        
         await HandleUploadResponse(uploadResponse);
         Dispose();
         return;
@@ -101,18 +146,25 @@ public partial class UploadFileTransfer : IDisposable
             => _selectedFile.OpenReadStream(_selectedFile.Size);
     }
 
-    private async Task SetProgressMessage(string message)
+    private async Task SetProgressMessageAsync(string message)
     {
         UploadStatusMessage = message;
         StateHasChanged();
         await Task.Delay(400);
     }
 
+    private void SetUploadPercentage(double percentage)
+    {
+        UploadStatusPercent = percentage;
+        InvokeAsync(StateHasChanged);
+    }
+    
     public void Dispose()
     {
         _selectedFile = null;
         EncryptionInProgress = false;
         _dropClass = string.Empty;
+        UploadStatusPercent = null;
         GC.SuppressFinalize(this);
     }
 }
